+++ api-contract/base/Contract.ts
@@ -2,7 +2,6 @@
 import type { ApiBase } from 'https://deno.land/x/polkadot/api/base/index.ts';
 import type { SubmittableExtrinsic } from 'https://deno.land/x/polkadot/api/submittable/types.ts';
 import type { ApiTypes, DecorateMethod } from 'https://deno.land/x/polkadot/api/types/index.ts';
-import type { Bytes } from 'https://deno.land/x/polkadot/types/mod.ts';
 import type { AccountId, ContractExecResult, EventRecord, Weight, WeightV2 } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
 import type { ISubmittableResult } from 'https://deno.land/x/polkadot/types/types/index.ts';
 import type { Abi } from '../Abi/index.ts';
@@ -112,9 +111,9 @@ export class Contract<ApiType extends ApiTypes> extends Base<ApiType> {
       // ContractEmitted is the current generation, ContractExecution is the previous generation
       new ContractSubmittableResult(result, applyOnEvent(result, ['ContractEmitted', 'ContractExecution'], (records: EventRecord[]) =>
         records
-          .map(({ event: { data: [, data] } }): DecodedEvent | null => {
+          .map((record): DecodedEvent | null => {
             try {
-              return this.abi.decodeEvent(data as Bytes);
+              return this.abi.decodeEvent(record);
             } catch (error) {
               l.error(`Unable to decode contract event: ${(error as Error).message}`);